hero page refactor and creating navbar component
